+++ src/Bundles/Exchange/HTML/HTMLImporter.php
@@ -2,14 +2,13 @@
 
 namespace PHPUnuhi\Bundles\Exchange\HTML;
 
-use PHPUnuhi\Bundles\Exchange\ImportInterface;
 use PHPUnuhi\Bundles\Exchange\ImportResult;
 use PHPUnuhi\Bundles\Storage\StorageInterface;
 use PHPUnuhi\Models\Translation\Translation;
 use PHPUnuhi\Models\Translation\TranslationSet;
 use SplFileObject;
 
-class HTMLImporter implements ImportInterface
+class HTMLImporter
 {
 
     /**