+++ src/Bundles/Translator/Fake/FakeTranslator.php
@@ -2,8 +2,8 @@
 
 namespace PHPUnuhi\Bundles\Translator\Fake;
 
-use PHPUnuhi\Bundles\Translator\CommandOption;
 use PHPUnuhi\Bundles\Translator\TranslatorInterface;
+use PHPUnuhi\Models\Command\CommandOption;
 
 class FakeTranslator implements TranslatorInterface
 {